Hotfix ETP-1618: Filter subprojects by deps
